# NL: Improve performance of `variable in place` queries