--- a/server/integration_tests/test_data/demo2_cities_feb2023/query_2/chart_config.json
+++ b/server/integration_tests/test_data/demo2_cities_feb2023/query_2/chart_config.json
@@ -23,59 +23,10 @@
               {
                 "tiles": [
                   {
-                    "comparisonPlaces": [
-                      "geoId/0677000"
-                    ],
-                    "statVarKey": [
-                      "Percent_Person_WithAllTeethLoss_multiple_place_bar_block",
-                      "Percent_Person_WithArthritis_multiple_place_bar_block",
-                      "Percent_Person_WithAsthma_multiple_place_bar_block",
-                      "Percent_Person_WithCancerExcludingSkinCancer_multiple_place_bar_block",
-                      "Percent_Person_WithChronicKidneyDisease_multiple_place_bar_block",
-                      "Percent_Person_WithChronicObstructivePulmonaryDisease_multiple_place_bar_block",
-                      "Percent_Person_WithCoronaryHeartDisease_multiple_place_bar_block",
-                      "Percent_Person_WithDiabetes_multiple_place_bar_block",
-                      "Percent_Person_WithHighBloodPressure_multiple_place_bar_block",
-                      "Percent_Person_WithHighCholesterol_multiple_place_bar_block",
-                      "Percent_Person_WithMentalHealthNotGood_multiple_place_bar_block",
-                      "Percent_Person_WithPhysicalHealthNotGood_multiple_place_bar_block",
-                      "Percent_Person_WithStroke_multiple_place_bar_block"
-                    ],
-                    "title": "Compared with Other Variables in Sunnyvale (${date})",
-                    "type": "BAR"
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            "columns": [
-              {
-                "tiles": [
-                  {
                     "statVarKey": [
                       "Percent_Person_Smoking"
                     ],
                     "title": "Percentage of Population That Smokes in Sunnyvale",
-                    "type": "LINE"
-                  }
-                ]
-              }
-            ]
-          },
-          {
-            "columns": [
-              {
-                "tiles": [
-                  {
-                    "statVarKey": [
-                      "Percent_Person_BingeDrinking",
-                      "Percent_Person_Obesity",
-                      "Percent_Person_PhysicalInactivity",
-                      "Percent_Person_SleepLessThan7Hours",
-                      "Percent_Person_Smoking"
-                    ],
-                    "title": "Compared with Other Variables in Sunnyvale",
                     "type": "LINE"
                   }
                 ]
@@ -106,6 +57,55 @@
                       "Percent_Person_WithChronicObstructivePulmonaryDisease"
                     ],
                     "title": "Chronic Obstructive Pulmonary Disease in Sunnyvale",
+                    "type": "LINE"
+                  }
+                ]
+              }
+            ]
+          },
+          {
+            "columns": [
+              {
+                "tiles": [
+                  {
+                    "comparisonPlaces": [
+                      "geoId/0677000"
+                    ],
+                    "statVarKey": [
+                      "Percent_Person_WithAllTeethLoss_multiple_place_bar_block",
+                      "Percent_Person_WithArthritis_multiple_place_bar_block",
+                      "Percent_Person_WithAsthma_multiple_place_bar_block",
+                      "Percent_Person_WithCancerExcludingSkinCancer_multiple_place_bar_block",
+                      "Percent_Person_WithChronicKidneyDisease_multiple_place_bar_block",
+                      "Percent_Person_WithChronicObstructivePulmonaryDisease_multiple_place_bar_block",
+                      "Percent_Person_WithCoronaryHeartDisease_multiple_place_bar_block",
+                      "Percent_Person_WithDiabetes_multiple_place_bar_block",
+                      "Percent_Person_WithHighBloodPressure_multiple_place_bar_block",
+                      "Percent_Person_WithHighCholesterol_multiple_place_bar_block",
+                      "Percent_Person_WithMentalHealthNotGood_multiple_place_bar_block",
+                      "Percent_Person_WithPhysicalHealthNotGood_multiple_place_bar_block",
+                      "Percent_Person_WithStroke_multiple_place_bar_block"
+                    ],
+                    "title": "Asthma compared with other variables in Sunnyvale (${date})",
+                    "type": "BAR"
+                  }
+                ]
+              }
+            ]
+          },
+          {
+            "columns": [
+              {
+                "tiles": [
+                  {
+                    "statVarKey": [
+                      "Percent_Person_BingeDrinking",
+                      "Percent_Person_Obesity",
+                      "Percent_Person_PhysicalInactivity",
+                      "Percent_Person_SleepLessThan7Hours",
+                      "Percent_Person_Smoking"
+                    ],
+                    "title": "Percentage of Population That Smokes compared with other variables in Sunnyvale",
                     "type": "LINE"
                   }
                 ]