… server/integration_tests/test_data/demo_fallback/query_3/chart_config.json
… server/integration_tests/test_data/demo_fallback/query_4/chart_config.json
… server/integration_tests/test_data/place_detection_e2e/query_3/chart_config.json
… server/integration_tests/test_data/screenshot/timeline/screenshot.json
… server/integration_tests/test_data/textbox_sample/query_1/chart_config.json
… server/tests/lib/nl/fulfiller_test.py